pipeline

+++ npm/mach/_scripts/postinstall/postinstall.mjs
@@ -1,8 +1,9 @@
 import * as fs from "node:fs";
 import * as path from "node:path";
-import * as http from "node:https";
 import * as url from "node:url";
 import * as child_process from "node:child_process";
+import { http_get, http_get_json } from './utils.mjs'
+import * as semver from "./vendor/semver/index.js";
 
 const __dirname = url.fileURLToPath(new URL(".", import.meta.url));
 
@@ -29,7 +30,7 @@ if (!ARCH || !OS) {
   process.exit(0)
 }
 
-const BIN_URL = `https://github.com/alshdavid/mach/releases/download/${BUILD_TAG}/mach-${OS}-${ARCH}.tar.gz`;
+const RELEASE_MANIFEST = "https://api.github.com/repos/alshdavid/mach/releases"
 const DOWNLOAD_TO = path.join(__dirname, "..", '..', "mach.tar.gz")
 
 if (process.platform === 'win32') {
@@ -40,43 +41,26 @@ if (process.platform === 'win32') {
 
 fs.rmSync(path.join(__dirname, "..", '..', "mach"), { force: true });
 fs.rmSync(path.join(__dirname, "..", '..', "mach.tar.gz"), { force: true });
+const github_manifest = await http_get_json(RELEASE_MANIFEST)
 
-function download_file_legacy(target_url, download_to) {
-  let buffer = undefined
-
-  function _download_file(target_url) {
-    return new Promise(res => {
-      const request = http.get(target_url, async (response) => {
-        const redirect = response.headers.location;
-
-        if (redirect) {
-          res(await _download_file(redirect))
-        } else {
-          response.on('end', () => res(buffer))
-          response.on('data', (chunk) => {
-            if (!buffer) {
-              buffer = chunk;
-            } else {
-              buffer = Buffer.concat([buffer, chunk]);
-            }
-          })
-        }
-      });
-      request.end()
-    })
+let bin_url = ''
+for (const gh_release of github_manifest) {
+  if (semver.satisfies(gh_release.tag_name, BUILD_TAG, { includePrerelease: true })) {
+    bin_url = `https://github.com/alshdavid/mach/releases/download/${gh_release.tag_name}/mach-${OS}-${ARCH}.tar.gz`
+    break
   }
-    
-  return _download_file(target_url)
+}
+if (!bin_url) {
+  console.error('Could not find Mach binary for specified version. Please compile from source')
+  process.exit(0)
 }
 
-const buffer = globalThis.fetch 
-  ? await fetch(BIN_URL).then(res => res.arrayBuffer()) 
-  : await download_file_legacy(BIN_URL)
+const buffer = await http_get(bin_url)
 
 fs.writeFileSync(DOWNLOAD_TO, Buffer.from(buffer));
 
 try {
-  child_process.execSync(`tar -xzf mach.tar.gz`, { cwd: path.resolve(__dirname,  '..'), stdio: 'inherit' })
+  child_process.execSync(`tar -xzf mach.tar.gz`, { cwd: path.resolve(__dirname,  '..', '..'), stdio: 'inherit' })
   fs.rmSync(DOWNLOAD_TO, { force: true })
 } catch (err) {
   console.error('Error: "tar" command is not installed')
